Fix rendering of lines longer than 2^16
Before things would be cast to u16 earlier than needed, which would
cause problems for insanely long lines (longer than 2^16 ~ 65 thousand)

+++ helix-view/src/editor.rs
@@ -558,7 +558,7 @@ impl Default for WhitespaceCharacters {
 pub struct IndentGuidesConfig {
     pub render: bool,
     pub character: char,
-    pub skip_levels: u16,
+    pub skip_levels: usize,
 }
 
 impl Default for IndentGuidesConfig {
